Refactor frame_sync.py into modular architecture
Major refactoring to improve code organization and maintainability:

**Problem:**
- frame_sync.py was 3,138 lines in a single monolithic file
- All sync modes, utilities, and helpers mixed together
- Difficult to navigate, test, and maintain
- Changes to one mode could accidentally affect others

**Solution:**
Split into clean modular architecture:

1. **frame_sync.py** (88 lines, was 3,138)
   - Now a thin orchestrator that re-exports from specialized modules
   - Maintains backward compatibility (no breaking changes)
   - 97% size reduction

2. **frame_utils.py** (1,274 lines)
   - Frame timing conversions (CFR and VFR)
   - VapourSynth indexing and frame extraction
   - Scene detection using PySceneDetect
   - Video property detection (FPS, interlacing, etc.)
   - Frame alignment validation

3. **checkpoint_selection.py** (87 lines)
   - Smart checkpoint selection for verification
   - Avoids OP/ED sequences, prefers dialogue

4. **sync_modes/** package (4 files, 1,869 lines total)
   - `time_based.py`: Raw delay sync without frame verification
   - `duration_align.py`: Duration-based alignment with hybrid verification
   - `correlation_frame_snap.py`: Audio correlation + scene detection
   - `subtitle_anchored_frame_snap.py`: Visual-only sync using subtitle positions

**Benefits:**
- ✅ Each mode is self-contained (easier to debug and test)
- ✅ Shared utilities in one place (no duplication)
- ✅ Changes to one mode don't affect others
- ✅ Clearer imports and dependencies
- ✅ Easier code review
- ✅ Backward compatible (all functions re-exported)
- ✅ Syntax validated (all files compile successfully)

**Total:** 3,318 lines across 8 files (vs. 3,138 in 1 file)
